refactor(file-source): use cxx shared enums for ResourceKind and FsErrorReason
Per review feedback (#164): the static_assert block was reinventing
what cxx already does for shared enums. Switch to the same pattern
the crate uses for MapLoadError / MapObserverCameraChangeMode.

- src/cpp/rust_file_source.h adds 'using ResourceKind = mbgl::Resource::Kind;'
  and 'using FsErrorReason = mbgl::Response::Error::Reason;' (mirrors the
  MapObserverCameraChangeMode trick for nested enums).
- src/renderer/bridge.rs declares both as #[namespace = mln::bridge]
  shared enums in the file_source bridge module. cxx codegen now
  validates name + discriminant alignment with the C++ aliases at
  compile time.
- RustFsResponse.error_reason changes from u8 with 0-as-sentinel to the
  bridged FsErrorReason; FsErrorReason::Success is the no-error sentinel
  (matches mbgl's convention that Reason starts at Success = 1).
- src/renderer/file_source.rs drops the hand-rolled #[repr(u8)] enums
  and the from_u8 mapper. The Rust callback signature is now
  Fn(&str, ResourceKind) -> FsResponse with ResourceKind as the typed
  bridged enum, so the callsite no longer needs a u8 -> enum match.
- src/cpp/rust_file_source.cpp drops the static_assert block and the
  static_cast<uint8_t>(resource.kind) at the bridge boundary; the kind
  flows through as the typed enum.
- Re-exports stay in place (crate::renderer::file_source::ResourceKind etc.)
  so downstream import paths are unchanged.

Test: tests/file_source_callback passes (0.35 s).
