+++ src/renderer/mod.rs
@@ -14,7 +14,8 @@ pub use bridge::{
     set_log_thread_enabled, Height, ScreenCoordinate, Size, Width, X, Y,
 };
 pub use builder::ImageRendererBuilder;
-pub use file_source::{FileSourceRequestCallback, FsErrorReason, FsResponse, ResourceKind};
+pub use bridge::file_source::{FsErrorReason, ResourceKind};
+pub use file_source::{FileSourceRequestCallback, FsResponse};
 pub use image_renderer::{Continuous, Image, ImageRenderer, RenderingError, Static, Tile};
 pub use map_observer::MapObserver;
 pub use resource_options::ResourceOptions;
